Move enrich channels tests to standalone class
